Fixing Have based tests
I quote:
Are recalculating the hash as part of the read progress update? What you should be providing in have is <bookmark_id>:<bookmark_hash>:<read_progress>:<read_time>. If you're updating read_progress and read_time, but not modifying bookmark_hash then it will not be returned. You should alter or modify bookmark_hash in some way in order to have the bookmarks received as part of the original /list request.
Hopefully that helps, but let me know if that is still not working for you or if there's anything else. Cheers!