Updated limit on list queries to get all as much data as possible from the service.
However, this isn't the promise of ALL the data; there is no API to get all the data that is in the remote service at this time (not even the official client does it).

Also, noted with this change is that likes are *super* noisey on sync, so left at 50 for the moment.

Additoinal change: added notification of each folder  as it syncs the book marks in that folder