Add support for "limits" in the sync & unit tests
Of note is the support for "hetrogenous" likes vs bookmarks.

E.g. first 25 bookmarks in a folder might not be enoug to cover the first 25 *liked* bookmarks. This means that you'd get errors when trying to sync that liked bookmark. This, it turned out, was a problem with my own personal collection.