Adding sort of the articles when displayed. This is done by inplace sort in the array of bookmarks returned from the DB.
One "faster" method that was investigated was to use IndexedDB indexes w/ openKeyCusor, but this doesn't work becaue the results are ordered by the key, and we want to *first* restrict the result set to the folder, and then sort the results on the bookmark_id. Because these are disparate fields, we're screwed, so lets do it the shitty way.