Refactoring App Object to allow re-used with upcoming share target support.
Specifically, now the app object is generic and handles the two signed in/out elements, along with switching between the two. Derived class knows how to create the "signed in experience", but all other things are handled wiht a shared SignedOut experience.

Removed need for exposing a shared instance of the app.

BOOM!