Refactored signing in to remove initialization being dependent on the UI "initializing" it.
Now the View Model does the init, but holds on being told the UI is ready before querying the first folder.

Also, tweaked a little login so that a sync is kicked off immediately for a "new" login. It's not kicked off if just starting the app (that ball of wax is for another day)