Fix issue where YouTube video would keep playing even after the article was dismissed.
This was due to:
a) Not cleaning shit up in the messenger that caused some shit to leak. Fixed now!
b) The face that the webview is native (and thus ref counted), but isn't ref counted to 0 until the GC runs, which as we all know w/ gc,is non-deterministic. Thus is can be a few seconds before the WebView is GC'd. Adding a navigate to "about:blank" to immediately stop it and then let the GC do it's thing