Added sign in / sign out animation to provide some continutity on the signing in / out of the app.
a) Added support to experiencehost to return the container when the view model is added.
b) Added extra property to signout to say if we're signinout from previously being signed in.
c) added support on authentication viewmodel to allow the consumer to ask for the spinner rto remain on screen once signed in.