Update supported VS version to 2017
This also brings in TypeScript 2.1, due to the way VS handles this.

There are some benefits here:
a) Better support for VS in the project files. Now it's possible to just add a file in the IDE
b) better language support for WinMD's: Now VS will generate the d.ts file from the WinMDs used by the SDK, so you don't need to update them manually. Yay!

However, with this, come some bugs:
a) https://github.com/Microsoft/TypeScript/issues/15322, which means shit is a little weird
b) Lots of warnings that are vomited for reasons I don't really understand.