Fix Move folder dialog loading
This change addresses an issue where the list of folders wasn't loading, when the popup was opened.

It's not 100% sure why this is happening -- or, more specifically, it's unclear where it got broken -- but it's probably because of the way the list view calculates it's size, which is a little weird because it's in the Flyout which has "deferred" layout. This means the listview is gonna go "Eh, no size" and not render anything.

This rejiggers it so it starts the show THEN loads the folder list.