Fixing error handling in the ShareTarget
if an error occured when sharing, the first error was handled gracefully. However, if you tried to share it a second time it would crash due to the way the share API didn't let you call various methods more than once.

This enlighted me to the fact that this statement applies to all the report* methods, and also theres somewhat of a state machine (don't call completed if you hadn't called start).

So, added some state to track what we have & haven't done.

Also, realized that in the error case we weren't indicating anything to the customer, so updated strings of save & the "description" text to say y'all can retry.