First pass at empty state
This change refactors some layout to provide some commonality of the "content" location in the tree for holding the content list and the empty state.

The empty state is a placeholder right now.

Of note, this *doesn't* handle the cases where the folder being viewed transitions from items->empty, nor from empty->items.