Fixing Sign-in Animation glitch
When signing in, there was an animation glitch where the entract animation would glitch to 50% and then animate, making it look pretty crappy.

What was the issue? Well, one could really have predicted this, but during load, we tried to set focus to the listviews first item by setting it's currentItem property. This moved the keyboard focus, which caused some scroller somewhere to scroll (likely a bug in edge).

So, rather than have it do this jiggery pokery immediately on loading, now we wait for a (newly added) event that indicates when the animation completes. _Then_ the focus is moved using the same mechanism.