Fixing telemetry issues in share target + sign out crasher
When sharing, the activation path is a little different (possible due to my own fault), because the telemetry hasn't restored it self yet. So, refactored the startup of the share target to be totally depenedent on the share activation event.

Additionally, while testing, found an intermittent crasher in signing out -- so attempted to handle it by defering setting of the saved settings over a timeout.