Refactored entrypoints for closing an artile
Before we hamfistedly pushed all the changes into a single `close` method that tried to handle all the different entry points based on stated.

Now, we say 'screw it', and push the decisions out to the direct handlers for those entry points (sane, right?)

These all get to make the correct decisions.

Motivation for making this change was that deleteing-while-fullscreen would delete the article but the viewer wouldn't dismiss.