Fixing click-click-click fidget behaviour in artcle
When the customer is clicking thought an article, they'll highlight/lose highlight, and this was toggling the toolbar. This is really distracting.

So, now we track that we had some selection, and now we don't -- we don't want to do anything then. Only when we no-select to no-select do we want to toggle the toolbar.