Fix issue with crashing after initial sign in
We were completing a promise twitce -- when we do the initial sign in, we don't wait for the full sync of articles to finish before dismissing the sign in screne. In a recent change all this completion was refactored and we missed this one case.

Now, if we're not waiting for downloads, don't try to complete the promise at the very end.