Move toolbar title into the article itself to make it "seamless"
This is step 1 of N to move the title of the article from being duplicated (article, and in the toolbar) to solely being in the article.

This is done by making it sticky when the toolbar is visible, and let it participate in the normal flow when the toolbar isn't visible.

This is incomplete visually, and the title/URL overflow the "underlay" backing. This will be addressed in a later change.