Fixed issue with scrollbars + pre text + wrapping it
Two issues:
a) Pre Text doesn't wrap: For code, thats kinda what you want. But for other text, not wrapping is a royal pain. So, risking it for now
b) When it wasn't wrapped, the scrollbar occluded the text, so adding a bottom padding on "pre" elements to allow room for the scrollbar.