Fix issue with images in long docs not having their width set
Some documents with lots of images don't have the their width correctly set, so bleed out of their intended space

This appears to be because, sometimes, when we ask for naturalWidth of the image it returns 0. Inspecting in the debugger returns the correct size. So, since if it's too large the clientWidth will ALSO be too wide, we check for both.