Fix issue when restoring article related to article with no progress
If an article has no progress locally (not just 0, but no update timestamp either), we were trying to say "update progress with a 0 timestamp", which the service rejected. Now set a low timestamp if we don't have one so we cna get the progress i it's changed, and avoid accidently updating it.