Handle Titlebar spacing changing for fullscreen/tablet mode
Sadly, there is still a delay on first launch -- likely because we're using an API that isn't supposed to be used by WWAs, and it says the titlbar isn't visible.

We _Tried_ to hide the UI until the first titlbar event, but that event isn't raised until _After_ the deferal for activation is complete, so bit of a catch 22.