Deleting an article wasn't propagating quickly enough
When we made a delete, we would make delete in the DB, but let the DB idle time actually do the sync.

However, this didn't happen quickly enough if you deleted + closed. So, added a little check that will cause an immediate sync if it's a delete book mark change.