Fixing issue with certain pages that contained iframes
The new custom scroller library had a bug with pages that contained iframes. Specifically, there was/is an assumption that if something is array like it's an array. With iFrames, window looks like that, and thus OverlayScrollbars breaks.

This introduces an explicit check for that case, and makes sure it's not thought to be Array like