Really fixing the blurry title in the article list
This appears to be a really subtle issue with rounding, and drop shadows. Specifically, we're using _two_ shadows to produce the backplate effect. Depsite these being on the same axis, one of them (the 'right' hand) is being rounded a little, so it moved up, leaving the extent of the left hand one showing through. This is weird, as it only shows at 100% scale.

Fix is three fold:
- Revert previous attempt to fix it
- Adjust the right hand blur to be 1px 'spread', which make sit look better on the right edge. This breaks the right edge for single-line titles so...
- ... introduce a box div thats about the right size, and is all kinds of offset to move around in the UI to _look_ like it's the same. This involves non-integral layout values, but it seems to work on all scale factors.

Additionally after all of this, 2nd-line text that was longer than the first would have it's top line peak though. Vertical-align with a value to attempt to fix that.