Add support for downloading an unsync'd article through deep link
This is a little bit of a hack since the service doesn't let you say 'give me bookmark 1234'. So, we use the 'updateProgress' API to update the progress with a _really_ old timestamp, which should, hopefully, leave the article unchanged, but give us enough info.

We then cram it in the orphaned folder, which on the next sync will delete the article. Sadly a sync _is_ started because the act of inserting the article into the DB for syncing, so it gets cleaned up. However, we've shown the article by this point, so it's OK (For now).

Additionally, refactored the 'download an article' 'cause it was the 3rd case of needing to do it.