Don't keeps splashscreen up if downloading and article.
When an article is already available locally, we can just open it, and hold the splash screen to allow the article to complete displaying. This presents a consistent, non-flickering/jumping UI.

However, if the article needs to be downloaded, it might take a while, and we can only hold the splash screen for a short period of time.

This change does a small refactor in the downloading case to return as soon as we try to download the article.