Add Fullscreen spinner control + consume in the deep linking experience
This adds a full screen spinner control that can be asked to display after a time out, and if canceled before the time out does not display. It also provides a simple message on the screen & cancel button in case you want to cancel it.

This change also integrates the dialog into the deeplinking experience.