Support a new link coming in while we're already trying to open one.
When a link is opening, it might take a while to open. This means that someone might click an inbound link a second time; this means that we need to cancel the first one and get on with the second one.

This change supports this by supporting cancellation of the promise that is attempting to display the article. With this, it can then kick off another one.