If an article is already being shown, hide it, and show the new one
This happens when one is viewing an article and an inbound link happens (Theres currently no other gesture to show an aritcle while one is already visible).

This adds a new static property, which holds the current article. When an article is displayed, it closes the current article if there is one.