Support opening the DB with a specific name
To support upcoming tests, this change allows you to open the DB with a specific name; this is in the hope to keep some tests isolated from some other tests.

This adds two fields to the `InstapaperDB` class (`_name`, `_version`), which are set via the initalize method.

Added tests to make sure things behave appropriately.