Adding basic 'What to read' class + test
This is the basic implementation + test. More to come.

In this case, this class is intended to support a way to simply get a grouped list of 'What to read' articles. The idea being that based on certain time stamps (progress, added), we can guess what the customer would want to read next.

The first use will be for the jump list in the taskbar & start menu.

In this change:
- `WhatToRead` class that queries the DB and decodes it's list directly from that by sorting in custom ways
- Tests for the above functionality, including generation of sample bookmarks

Additional payload to expose a couple of interfaces for testing.