Respect user removal from the Shell jumplist
When the customer 'removed' an item using the taskbar gesture, we should make sure we don't add it.

This is a _mostly_ complete implementation -- specifically, if they _remove_ the item, and then it ages out (e.g. they add some more articles, and it was an old one), we will remove it from the DB, removing the removed from jumplist flag that we added to it.