Again: Fixing issue with animation on sign in, where it glitched
Whats going on? So, when we sign in, the UI is transformed off screen. _However_ at the same time as soon as we have some content, we would focus the first element. Edge has a terrible issue that when something is transformed off screen, and you _focus it_, the browser will attempt to scroll the elements into view, and... then just leaves it there.

So, we shift the moment when set the focus till after we've animated things into place.

While investigating, it also became clear the first sync wait wasn't correctly ignoring the article download, and was waiting for it -- e.g. the article & images download -- to complete.

So, went and fixed that.