Add Correct displaying of whats new banner
Whats new banner now only shows it if it was not previously displayed for some opaque identifier (E.g. if it's different, show it -- no matter how long it was).

Additionally, this behaviour is only triggered if we're automatically signing in (e.g. someone had previously signed in during a previous launch), so we don't show it on first run.

Additional: removed installID 'persistence' since that is now managed by the telemetry library itself.