Remove usage of `asyncTest` and switchtc to `assert.async()`
QUnit has deprecated `asyncTest`, and switch to using the (passed to every test) `assert` object for the `async` method.

As part of this, finally understood that the new model for QUnit is to pass an assert object to each test (localizing the exposed surface), so updated the `promiseTest` to call this, and the typescript defn. for these things to make it optional (for now).

Additionally, removed some globals from the QUnit implementation to help catch some of these failures.