Remove Global `QUnit.assert` usage
This API was deprecated and moved into a parameter that's passed to each test. This updates our usage so that all tests functions now take an `assert` parameter.

Of note, a number of places have had to be updated to 'thread' the assert class all the way down. This is ugly, but now is not the time to embark on a big clean up of that.