Reformatted tests to be 'inline' rather than explicitly functions
This is to help with the migration to a BDD interface, rather than the QUnit interface. the convention there is to have it inline with the description. There are additional challenges for _some_ migrations, but this is a good starting point.

Additional:
Removed 'dependent upon' 'cause it was the only one using it in the test project.