Converted InstapaperDB Store tests to be BDD
Note, this changes core utility functions, and probably breaks other, non-converted tests.

Note, the change for `deleteDb` to be `tidyDb`, and a `.bind(null, null)`. This is because Mocha inspects the argument count on a function, and says that if the test function accepts > 0 parameters, it can only be completed by calling the parameter value it supplies at test time.

Since `deleteDb` is a helper with an optional parameter this buggered things up right n proper when used explicitly as the test itself.