Moving the static fields from `InstapaperDB` to be enums in the `Codevoid.Storyvoid` namespace
When originally created, I intentionally created `InstapaperDB` as a class w/ a bunch of statics. In hindsight that was a bad way to express things; but we didn't have `enum`'s back in the day.

This change adjusts all consumers as best can be found at this time -- all tests impacted passed.