Remove `WinJS.Promise` references from return types + Remove Progress support on `Signal`
The removal of progress reporting on `Signal` is that progress is not a concept in native promises, and no-where was using it, so... LATERS! This required a small tweak in db.js (which will eventually go away) to use raw `WinJS.Promise` to continue the same model it previously had for progress.

Everything else is _type_ changes, and shouldn't have any semantic differences. There were _liberal_ sprinklings of any to make types sort themselves out. This is all a prelude to moving to `async` / `await` syntax, which should enable better typing, so hopefully are just a stop gap.