Fixing user-reported crash on closing an article
Reported by a user is a crash that seems to occur in `end()`, when
cleaning up the events. This was guarded by a check to see if the
session was setup, but it looks like it's failing an odd & different
point during setup, so we were passing the check on clean up
despite having incomplete state.

To address this, moved the setting of successful creation later
and updated the clean up to be more resiliant.