Make opening solution & building work with no configuration
To make it easyier to open-and-go, it would be be nice if:
a) The default target was one your device was likely to support
b) The app, and not tests, was the default selected project

This change renames the `x64` flavour in the solution to be `AMD64` which will sort first.

It also reorder the reference to the app project to make sure it's default selected in a new clone.

Lastly: Remove `AnyCPU` since we don't support that 'cause of the native code